o Updated to create the 'sessionDirectory' and to fail the build when creating that directory fails.
git-svn-id: http://svn.jomc.org/svnroot/jomc/jomc/trunk/maven-jomc-plugin@2414 aee2edba-5628-49ba-b8b8-0e9d4ac63d57

+++ src/main/java/org/jomc/mojo/AbstractAttachMojo.java
@@ -207,6 +207,12 @@ public abstract class AbstractAttachMojo extends AbstractMojo
                     {
                         this.getLog().warn( getMessage( "failedDeletingFile", attachment.getAbsolutePath() ) );
                     }
+                    if ( !attachment.getParentFile().exists() && !attachment.getParentFile().mkdirs() )
+                    {
+                        throw new MojoExecutionException( getMessage( "failedCreatingDirectory",
+                                                                      attachment.getParentFile().getAbsolutePath() ) );
+
+                    }
 
                     FileUtils.copyFile( this.getArtifactFile(), attachment );
                     this.getMavenProjectHelper().attachArtifact( this.getMavenProject(), this.getArtifactType(),
